# Use array instead of linked list for rbs  location's child