--- a/ext/rbs_extension/location.c
+++ b/ext/rbs_extension/location.c
@@ -2,54 +2,6 @@
 
 VALUE RBS_Location;
 
-rbs_loc_list *rbs_loc_list_add(rbs_loc_list *list, const ID name, const range r) {
-  rbs_loc_list *new = malloc(sizeof(rbs_loc_list));
-  new->next = list;
-  new->name = name;
-  new->rg = r;
-  return new;
-}
-
-rbs_loc_list *rbs_loc_list_dup(rbs_loc_list *list) {
-  if (list) {
-    return rbs_loc_list_add(rbs_loc_list_dup(list->next), list->name, list->rg);
-  } else {
-    return NULL;
-  }
-}
-
-void rbs_loc_list_free(rbs_loc_list *list) {
-  while (list) {
-    rbs_loc_list *next = list->next;
-    free(list);
-    list = next;
-  }
-}
-
-bool rbs_loc_list_find(const rbs_loc_list *list, ID name, range *rg) {
-  while (list) {
-    if (list->name == name) {
-      *rg = list->rg;
-      return true;
-    }
-
-    list = list->next;
-  }
-
-  return false;
-}
-
-size_t rbs_loc_list_size(const rbs_loc_list *list) {
-  size_t size = 0;
-
-  while (list) {
-    size += 1;
-    list = list->next;
-  }
-
-  return size;
-}
-
 position rbs_loc_position(int char_pos) {
   position pos = { 0, char_pos, -1, -1 };
   return pos;
@@ -60,24 +12,52 @@ position rbs_loc_position3(int char_pos, int line, int column) {
   return pos;
 }
 
+void rbs_loc_alloc_children(rbs_loc *loc, unsigned short size) {
+  size_t s = sizeof(rbs_loc_list) + sizeof(rbs_loc_entry) * size;
+  loc->list = malloc(s);
+
+  loc->list->len = 0;
+  loc->list->required_p = 0;
+  loc->list->cap = size;
+}
+
+void rbs_loc_alloc_adding_children(rbs_loc *loc) {
+  if (loc->list == NULL) {
+    rbs_loc_alloc_children(loc, 1);
+  } else {
+    if (loc->list->len == loc->list->cap) {
+      size_t s = sizeof(rbs_loc_list) + sizeof(rbs_loc_entry) * (++loc->list->cap);
+      loc->list = realloc(loc->list, s);
+    }
+  }
+}
+
 void rbs_loc_add_required_child(rbs_loc *loc, ID name, range r) {
-  loc->requireds = rbs_loc_list_add(loc->requireds, name, r);
+  rbs_loc_alloc_adding_children(loc);
+
+  unsigned short i = loc->list->len++;
+  loc->list->entries[i].name = name;
+  loc->list->entries[i].rg = r;
+
+  loc->list->required_p |= 1 << i;
 }
 
 void rbs_loc_add_optional_child(rbs_loc *loc, ID name, range r) {
-  loc->optionals = rbs_loc_list_add(loc->optionals, name, r);
+  rbs_loc_alloc_adding_children(loc);
+
+  unsigned short i = loc->list->len++;
+  loc->list->entries[i].name = name;
+  loc->list->entries[i].rg = r;
 }
 
 void rbs_loc_init(rbs_loc *loc, VALUE buffer, range rg) {
   loc->buffer = buffer;
   loc->rg = rg;
-  loc->optionals = NULL;
-  loc->requireds = NULL;
+  loc->list = NULL;
 }
 
 void rbs_loc_free(rbs_loc *loc) {
-  rbs_loc_list_free(loc->optionals);
-  rbs_loc_list_free(loc->requireds);
+  free(loc->list);
   ruby_xfree(loc);
 }
 
@@ -89,7 +69,7 @@ static void rbs_loc_mark(void *ptr)
 
 static size_t rbs_loc_memsize(const void *ptr) {
   const rbs_loc *loc = ptr;
-  return sizeof(*loc) + (rbs_loc_list_size(loc->optionals) + rbs_loc_list_size(loc->requireds)) * sizeof(rbs_loc_list);
+  return sizeof(rbs_loc) + sizeof(rbs_loc_list) + sizeof(rbs_loc_entry) * loc->list->cap;
 }
 
 static rb_data_type_t location_type = {
@@ -130,8 +110,7 @@ static VALUE location_initialize_copy(VALUE self, VALUE other) {
 
   self_loc->buffer = other_loc->buffer;
   self_loc->rg = other_loc->rg;
-  self_loc->requireds = rbs_loc_list_dup(other_loc->requireds);
-  self_loc->optionals = rbs_loc_list_dup(other_loc->optionals);
+  rbs_loc_alloc_children(self_loc, other_loc->list->cap);
 
   return Qnil;
 }
@@ -177,6 +156,7 @@ static VALUE location_end_loc(VALUE self) {
   }
 }
 
+// TODO: Check if we need to allocate more space
 static VALUE location_add_required_child(VALUE self, VALUE name, VALUE start, VALUE end) {
   rbs_loc *loc = rbs_check_location(self);
 
@@ -189,6 +169,7 @@ static VALUE location_add_required_child(VALUE self, VALUE name, VALUE start, VA
   return Qnil;
 }
 
+// TODO: Check if we need to allocate more space
 static VALUE location_add_optional_child(VALUE self, VALUE name, VALUE start, VALUE end) {
   rbs_loc *loc = rbs_check_location(self);
 
@@ -221,18 +202,17 @@ VALUE rbs_new_location(VALUE buffer, range rg) {
 static VALUE location_aref(VALUE self, VALUE name) {
   rbs_loc *loc = rbs_check_location(self);
 
-  range result;
   ID id = SYM2ID(name);
 
-  if (rbs_loc_list_find(loc->requireds, id, &result)) {
-    return rbs_new_location(loc->buffer, result);
-  }
+  for (unsigned short i = 0; i < loc->list->len; i++) {
+    if (loc->list->entries[i].name == id) {
+      range result = loc->list->entries[i].rg;
 
-  if (rbs_loc_list_find(loc->optionals, id, &result)) {
-    if (null_range_p(result)) {
-      return Qnil;
-    } else {
-      return rbs_new_location(loc->buffer, result);
+      if (!(loc->list->required_p & (1 << i)) && null_range_p(result)) {
+        return Qnil;
+      } else {
+        return rbs_new_location(loc->buffer, result);
+      }
     }
   }
 
@@ -244,11 +224,12 @@ static VALUE location_optional_keys(VALUE self) {
   VALUE keys = rb_ary_new();
 
   rbs_loc *loc = rbs_check_location(self);
-  rbs_loc_list *list = loc->optionals;
+  rbs_loc_list list = *loc->list;
 
-  while (list) {
-    rb_ary_push(keys, ID2SYM(list->name));
-    list = list->next;
+  for (unsigned short i = 0; i < list.len; i++) {
+    if (!(list.required_p & (1 << i))) {
+      rb_ary_push(keys, ID2SYM(list.entries[i].name));
+    }
   }
 
   return keys;
@@ -258,11 +239,12 @@ static VALUE location_required_keys(VALUE self) {
   VALUE keys = rb_ary_new();
 
   rbs_loc *loc = rbs_check_location(self);
-  rbs_loc_list *list = loc->requireds;
+  rbs_loc_list list = *loc->list;
 
-  while (list) {
-    rb_ary_push(keys, ID2SYM(list->name));
-    list = list->next;
+  for (unsigned short i = 0; i < list.len; i++) {
+    if (list.required_p & (1 << i)) {
+      rb_ary_push(keys, ID2SYM(list.entries[i].name));
+    }
   }
 
   return keys;