--- a/test/rbs/location_test.rb
+++ b/test/rbs/location_test.rb
@@ -5,11 +5,6 @@ class RBS::LocationTest < Test::Unit::TestCase
   Location = RBS::Location
 
   def test_location_source
-    buffer = Buffer.new(name: Pathname("foo.rbs"), content: <<-CONTENT)
-123
-abc
-    CONTENT
-
     Location.new(buffer, 0, 4).yield_self do |location|
       assert_equal 0, location.start_pos
       assert_equal 4, location.end_pos
@@ -30,11 +25,6 @@ abc
   end
 
   def test_location_child
-    buffer = Buffer.new(name: Pathname("foo.rbs"), content: <<-CONTENT)
-123
-abc
-    CONTENT
-
     Location.new(buffer, 0, 8).yield_self do |location|
       location.add_optional_child(:num, 0...2)
       location.add_optional_child(:hira, nil)
@@ -47,5 +37,51 @@ abc
       assert_equal [:num, :hira].sort, location.each_optional_key.to_a.sort
       assert_equal [:alpha], location.each_required_key.to_a
     end
+  end
+
+  def test_location_initialize_copy
+    loc = Location.new(buffer, 0, 8)
+    loc.add_optional_child(:num, 0...2)
+    loc.add_required_child(:alpha, 4...7)
+    assert_equal loc, loc.dup
+    assert_equal loc[:num], loc.dup[:num]
+    assert_equal loc[:alpha], loc.dup[:alpha]
+  end
+
+  def test_location_aref
+    loc_without_child = Location.new(buffer, 0, 8)
+    assert_raise RuntimeError do
+      loc_without_child[:not_exist]
+    end
+    loc = Location.new(buffer, 0, 8)
+    loc.add_optional_child(:num, 0...2)
+    loc.add_required_child(:alpha, 4...7)
+    assert_equal "12", loc[:num].source
+    assert_equal "abc", loc[:alpha].source
+  end
+
+  def test_location_start_pos
+    loc = Location.new(buffer, 0, 8)
+    assert_equal 0, loc.start_pos
+  end
+
+  def test_location_end_pos
+    loc = Location.new(buffer, 0, 8)
+    assert_equal 8, loc.end_pos
+  end
+
+  def test_location_to_s
+    loc = Location.new(buffer, 0, 7)
+    assert_equal "foo.rbs:1:0...2:3", loc.to_s
+  end
+
+  private
+
+  def buffer(content: nil)
+    content ||= <<~CONTENT
+      123
+      abc
+    CONTENT
+    Buffer.new(name: Pathname("foo.rbs"), content: content)
   end
 end